Split section management into its own file
