+++ src/asm/asmy.y
@@ -22,8 +22,8 @@
 #include "asm/lexer.h"
 #include "asm/main.h"
 #include "asm/mymath.h"
-#include "asm/output.h"
 #include "asm/rpn.h"
+#include "asm/section.h"
 #include "asm/symbol.h"
 #include "asm/util.h"
 #include "asm/warning.h"
